userRegister

+++ application/models/UserManageModel.php
@@ -7,31 +7,42 @@
 			$password = md5($pass);
 			$query=$this->db->query("SELECT * FROM user WHERE (username='$username' OR email='$username') AND password='$password'");
 			if($query->num_rows()==1){
-				$this->load->helper('cookie');
-				if($this->input->post('remember')){
-					$cookie1=array('name'=>'remember_me_user','value'=>$username,'expire'=>time()+50400,'path'=>'/');
-					$cookie2=array('name'=>'remember_me_pass','value'=>$pass,'expire'=>time()+50400,'path'=>'/');
-					set_cookie($cookie1);
-					set_cookie($cookie2);
-				}else{
-					$value1=get_cookie('remember_me_user');
-					$value2=get_cookie('remember_me_pass');
-            		if($value1 && $value2){
-            			
-			            if($value1==$username){
-			            	delete_cookie('remember_me_user');
-			            	delete_cookie('remember_me_pass');
-			            }
-            		}
-				}
-
 				$row=$query->row();
-				$newdata=array('username'=>$row->username,'email'=>$row->email);
-				$this->session->set_userdata($newdata);
-				redirect('','location');
-				return true;
+				$verify=$row->emailVerify;
+				
+				if($verify=='yes' || $verify=='standby'){
+					$this->load->helper('cookie');
+					if($this->input->post('remember')){
+						$cookie1=array('name'=>'remember_me_user','value'=>$username,'expire'=>time()+50400,'path'=>'/');
+						$cookie2=array('name'=>'remember_me_pass','value'=>$pass,'expire'=>time()+50400,'path'=>'/');
+						set_cookie($cookie1);
+						set_cookie($cookie2);
+					}else{
+						$value1=get_cookie('remember_me_user');
+						$value2=get_cookie('remember_me_pass');
+	            		if($value1 && $value2){
+	            			
+				            if($value1==$username){
+				            	delete_cookie('remember_me_user');
+				            	delete_cookie('remember_me_pass');
+				            }
+	            		}
+					}
+
+					$row=$query->row();
+					$newdata=array('username'=>$row->username,'email'=>$row->email);
+					$this->session->set_userdata($newdata);
+					if($verify=='yes'){
+						redirect('','location');
+					}else{
+						$query=$this->db->query("UPDATE user SET emailVerify='yes' WHERE username='$username'");
+						return 1;
+					}
+				}else{
+					return 2;
+				}
 			}else{
-				return false;
+				return 3;
 			}
 		}
 
@@ -48,14 +59,14 @@
 			return $query->result_array();
 		}
 
-		public function addNewUser(){
+		public function addNewUser($verifycode){
 			$username=$this->input->post('username');
 			$email=$this->input->post('email');
 			$password=$this->input->post('password');
-			$query=$this->db->query("INSERT INTO user(username,email,password) VALUES('$username','$email','$password')");
+			$query=$this->db->query("INSERT INTO user(username,email,password,emailVerify) VALUES('$username','$email','$password','$verifycode')");
 			if($query){
-				$newdata=array('username'=>$username,'email'=>$email);
-				$this->session->set_userdata($newdata);
+				//$newdata=array('username'=>$username,'email'=>$email);
+				//$this->session->set_userdata($newdata);
 				return true;
 			}else{
 				return false;
@@ -106,6 +117,38 @@
 			if($query2){
 				$newdata=array('username'=>$username,'email'=>$email);
 				$this->session->set_userdata($newdata);
+				return true;
+			}else{
+				return false;
+			}
+		}
+
+		public function getEmailVerifycode($username){
+			$query=$this->db->query("SELECT emailVerify FROM user WHERE username='$username'");
+			$row=$query->row();
+			return $row->emailVerify;
+		}
+
+		public function setStandby($username){
+			$query=$this->db->query("UPDATE user SET emailVerify='standby' WHERE username='$username'");
+		}
+
+		public function setUserDetails(){
+			$username=$this->session->userdata('username'); 
+			$fn=$this->input->post('firstname');
+			$ln=$this->input->post('lastname');
+			$nic=$this->input->post('nic');
+			$con=$this->input->post('contact');
+			$ad1=$this->input->post('address1');
+			$ad2=$this->input->post('address2');
+			$ad3=$this->input->post('address3');
+			$bn=$this->input->post('bankname');
+			$holder=$this->input->post('holdername');
+			$accnumber=$this->input->post('accnumber');
+			
+			$query=$this->db->query("UPDATE user SET firstname='$fn',lastname='$ln',nic='$nic',contact='$con',address1='$ad1',address2='$ad2',address3='$ad3',bank='$bn',holdername='$holder',accnum='$accnumber' WHERE username='$username'");
+			
+			if($query){
 				return true;
 			}else{
 				return false;
